+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+- Added **where-is** command, bound to **C-h w**, to look up what keys a command
+  is bound to, if any.
 - Added **describe-key-briefly** command, bound to **C-h c**, as a less verbose
   alternative to **describe-key**.
 - Added **print-buffer** command, bound to **C-x p** to print the current